finishing test, add possibility to use to_owned on TextualToken
