+++ src/token/tag.rs
@@ -1,11 +1,12 @@
 use std::default::Default;
+use std::clone::Clone;
 use crate::parser::{list, heading};
 use crate::parser::operator::bytes;
 
 // Minimum character length
 const MIN_CHAR_LENGTH: usize = 2;
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Clone)]
 pub enum TagOperator {
     Heading,
     UnorderedList,
@@ -17,7 +18,7 @@ impl Default for TagOperator {
     fn default() -> Self { TagOperator::Heading }
 }
 
-#[derive(Debug, Default)]
+#[derive(Debug, Default, Clone)]
 pub struct TagToken { 
     pub line: usize,
     pub content: String,
@@ -25,7 +26,7 @@ pub struct TagToken {
     pub metas: Option<TagMeta>
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct TagMeta {
     pub heading_kind: heading::HeadingLevel
 }
